refactor: move Dataset classes and augmentation from models to data
