+++ simba/core/data/datasets/__init__.py
@@ -1,0 +1,9 @@
+"""PyTorch Dataset classes for SIMBA."""
+
+from simba.core.data.datasets.encoder_dataset import CustomDatasetEncoder
+from simba.core.data.datasets.multitask_dataset import CustomDatasetMultitasking
+
+__all__ = [
+    "CustomDatasetEncoder",
+    "CustomDatasetMultitasking",
+]
