+++ simba/core/data/datasets/encoder_dataset.py
@@ -1,6 +1,6 @@
 from torch.utils.data import Dataset
 
-from simba.core.models.transformers.augmentation import Augmentation
+from simba.core.data.augmentation import Augmentation
 
 
 class CustomDatasetEncoder(Dataset):
